+++ x/emissions/keeper/migrator.go
@@ -6,6 +6,7 @@ import (
 	"github.com/zeta-chain/node/x/emissions/exported"
 	v5 "github.com/zeta-chain/node/x/emissions/migrations/v5"
 	v6 "github.com/zeta-chain/node/x/emissions/migrations/v6"
+	v7 "github.com/zeta-chain/node/x/emissions/migrations/v7"
 )
 
 // Migrator is a struct for handling in-place store migrations.
@@ -30,4 +31,9 @@ func (m Migrator) Migrate4to5(ctx sdk.Context) error {
 // Migrate5to6 migrates the store from consensus version 5 to 6
 func (m Migrator) Migrate5to6(ctx sdk.Context) error {
 	return v6.MigrateStore(ctx, m.keeper)
+}
+
+// Migrate6to7 migrates the store from consensus version 6 to 7
+func (m Migrator) Migrate6to7(ctx sdk.Context) error {
+	return v7.MigrateStore(ctx, m.keeper)
 }